move optionsFromArray to story utils
